ID-45: tidy up
