+++ service-api/docker-compose.env
@@ -1,0 +1,5 @@
+AWS_ACCESS_KEY_ID=localstack
+AWS_SECRET_ACCESS_KEY=localstack
+AWS_REGION=eu-west-1
+PAPER_ID_BACK_AWS_ENDPOINT=http://localstack:4566
+PAPER_ID_BACK_AWS_DEBUG=false
