+++ service-api/module/Application/src/Fixtures/Data/sampleData.json
@@ -16,6 +16,7 @@
       "PW M-1234-ABCD-AAAA"
     ],
     "verify_method": "passport",
+    "id_number": "1483792528",
     "role": "donar",
     "id_passed": 1,
     "kbq_passed": 1
@@ -37,6 +38,7 @@
       "PW M-1234-ABCD-AAFA"
     ],
     "verify_method": "passport",
+    "id_number": "1483532528",
     "role": "donar",
     "id_passed": 1,
     "kbq_passed": 0
@@ -58,6 +60,7 @@
       "PW M-1234-ABCD-AYUA"
     ],
     "verify_method": "dvla",
+    "id_number": "HA1483fs528",
     "role": "donar",
     "id_passed": 1,
     "kbq_passed": 1
@@ -79,12 +82,13 @@
       "PW M-1234-ABCD-AAAA"
     ],
     "verify_method": "dwp",
+    "id_number": "14AH387sdfj",
     "role": "donar",
     "id_passed": 1,
     "kbq_passed": 0
   },
   {
-    "id": 4,
+    "id": 5,
     "name": "James Brown",
     "dob": "1968-08-01",
     "address": {
@@ -100,6 +104,7 @@
       "PW M-1234-ABCD-AAAA"
     ],
     "verify_method": "dwp",
+    "id_number": "1483692528",
     "role": "voucher",
     "id_passed": 1,
     "kbq_passed": 0
